IPRT: strformat-vbox.cpp: More 64-bit MSC warnings.
svn:sync-xref-src-repo-rev: r39117

+++ src/VBox/Runtime/VBox/strformat-vbox.cpp
@@ -1,4 +1,4 @@
-/* $Id: strformat-vbox.cpp 11413 2008-08-14 08:03:03Z klaus.espenlaub@oracle.com $ */
+/* $Id: strformat-vbox.cpp 14061 2008-11-10 23:12:54Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - VBox String Formatter extensions.
  */
@@ -172,8 +172,8 @@ size_t rtstrFormatVBox(PFNRTSTROUTPUT pfnOutput, void *pvArgOutput, const char *
                                  */
                                 case 'd':
                                 {
-                                    int cch = 0;
-                                    int off = 0;
+                                    size_t  cch = 0;
+                                    int     off = 0;
 
                                     if (cchPrecision <= 0)
                                         cchPrecision = 16;
@@ -210,7 +210,7 @@ size_t rtstrFormatVBox(PFNRTSTROUTPUT pfnOutput, void *pvArgOutput, const char *
                                 {
                                     if (cchWidth-- > 0)
                                     {
-                                        int cch = RTStrFormat(pfnOutput, pvArgOutput, NULL, 0, "%02x", *pu8++);
+                                        size_t cch = RTStrFormat(pfnOutput, pvArgOutput, NULL, 0, "%02x", *pu8++);
                                         for (; cchWidth > 0; cchWidth--, pu8++)
                                             cch += RTStrFormat(pfnOutput, pvArgOutput, NULL, 0, " %02x", *pu8);
                                         return cch;
